+++ docs/examples/layout/row_styling.py
@@ -62,4 +62,4 @@ app.layout = html.Div(
 )
 
 if __name__ == "__main__":
-    app.run_server(debug=False)
+    app.run_server(debug=True)