+++ src/lib/fragments/AgGrid.react.js
@@ -337,7 +337,10 @@ export default class DashAgGrid extends Component {
             }
 
             if ('cellStyle' in colOut) {
-                if (Object.keys(colOut.cellStyle).includes('styleConditions')) {
+                if (
+                    Object.keys(colOut.cellStyle).includes('styleConditions') ||
+                    Object.keys(colOut.cellStyle).includes('defaultStyle')
+                ) {
                     const cellStyle = JSON.parse(
                         JSON.stringify(colOut.cellStyle)
                     );
@@ -597,6 +600,7 @@ export default class DashAgGrid extends Component {
         });
 
         this.updateColumnWidths();
+        this.updateColumnState();
 
         if (this.state.rowTransaction) {
             this.state.rowTransaction.map((data) =>